Create heap sort.cpp
C++ Heap Sort Heapsort is similar to the selection sort, where we first find the maximum element and place a maximum element at the end. We repeat the same process for the remaining item. In this, the largest element among the values is selected and then placed at the end of the value list (in case of ascending order sorting).
What is a Binary Heap
The complete binary tree is a binary tree in which every level, except possibly the last, is filled, and all nodes are as far left as possible.

The Binary Heap is a Complete Binary Tree where elements are stored in a unique order such that the value in the parent node is greater(or smaller) than the values in its two children nodes.

The former is called max heap, and the latter is called min-heap. The binary tree or array can represent the heap.